fix test fibonacci

+++ include/fibonacci.h
@@ -2,7 +2,7 @@
 #define FIBONACCI_H
 #include <stdint.h>
 
-uint32_t fibonacci_calculator(int n);
+uint64_t fibonacci_calculator(int n);
 
 #endif
